fix: resolve TypeScript errors and update imports for Apple design system
- Remove duplicate health controller, use enhanced version as main
- Fix TypeScript type errors with proper type annotations
- Remove framer-motion components, use react-spring versions
- Update imports to use magicui components from API
- Add missing createPaymentIntent function
- Fix Supabase SSR client creation in auth callback
- Remove 'any' types, use proper TypeScript types throughout
