+++ apps/frontend/src/app/auth/callback/route.ts
@@ -1,6 +1,8 @@
 import { NextResponse } from 'next/server'
 import { cookies } from 'next/headers'
-import { createServerSupabaseClient, logger } from '@repo/shared'
+import { createServerClient } from '@supabase/ssr'
+import { logger } from '@repo/shared'
+import type { Database } from '@repo/shared'
 
 export async function GET(request: Request) {
 	const { searchParams, origin } = new URL(request.url)
@@ -14,8 +16,29 @@ export async function GET(request: Request) {
 
 	if (code) {
 		try {
-			const cookieStore = cookies()
-			const supabase = createServerSupabaseClient(cookieStore)
+			const cookieStore = await cookies()
+			const supabase = createServerClient<Database>(
+				process.env.NEXT_PUBLIC_SUPABASE_URL!,
+				process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+				{
+					cookies: {
+						getAll() {
+							return cookieStore.getAll()
+						},
+						setAll(cookiesToSet) {
+							try {
+								cookiesToSet.forEach(({ name, value, options }) =>
+									cookieStore.set(name, value, options)
+								)
+							} catch {
+								// The `setAll` method was called from a Server Component.
+								// This can be ignored if you have middleware refreshing
+								// user sessions.
+							}
+						}
+					}
+				}
+			)
 			const { data, error } = await supabase.auth.exchangeCodeForSession(code)
 
 			if (error) {
