+++ apps/frontend/src/components/pricing/checkout/checkout-loading.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { MagicCard } from '@/components/ui/magic-card'
+import { MagicCard } from '@/components/magicui/magic-card'
 import { Loader } from '@/components/ui/loader'
 import { Lock } from 'lucide-react'
 import { cn } from '@/lib/utils'
