+++ apps/frontend/src/components/pricing/checkout/checkout-success.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { MagicCard } from '@/components/ui/magic-card'
+import { MagicCard } from '@/components/magicui/magic-card'
 import { CheckCircle2 } from 'lucide-react'
 import { cn } from '@/lib/utils'
 import { animated } from '@react-spring/web'
