+++ apps/frontend/src/components/pricing/checkout/error-banner.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Button } from '@/components/ui/button'
-import { MagicCard } from '@/components/ui/magic-card'
+import { MagicCard } from '@/components/magicui/magic-card'
 import { cn } from '@/lib/utils'
 import { animated } from '@react-spring/web'
 import { AlertTriangle, RefreshCw } from 'lucide-react'
