+++ apps/frontend/src/components/pricing/checkout/loading-overlay.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Loader } from '@/components/ui/loader'
-import { MagicCard } from '@/components/ui/magic-card'
+import { MagicCard } from '@/components/magicui/magic-card'
 import { cn } from '@/lib/utils'
 import { Lock } from 'lucide-react'
