+++ apps/frontend/src/components/pricing/checkout/use-checkout.ts
@@ -118,6 +118,7 @@ export function useCheckout({
 		status,
 		error,
 		confirmPayment,
-		reset
+		reset,
+		formatAmount
 	}
 }
